+++ exonum-java-binding/core/rust/src/storage/mod.rs
@@ -15,6 +15,7 @@
 pub use self::access::Java_com_exonum_binding_core_storage_database_Accesses_nativeFree;
 pub(crate) use self::access::{into_erased_access, EjbAccessExt};
 pub use self::blockchain::*;
+pub use self::blockchain_data::*;
 pub use self::key_set_index::*;
 pub use self::list_index::*;
 pub use self::map_index::*;
@@ -25,12 +26,13 @@ pub use self::proof_list_index::*;
 pub use self::proof_map_index::*;
 pub use self::proof_map_index_next::*;
 pub use self::raw_proof_map_index::*;
-pub use self::readonly_fork::*;
+pub use self::readonly_erased_access::*;
 pub use self::temporarydb::*;
 pub use self::value_set_index::*;
 
 mod access;
 mod blockchain;
+mod blockchain_data;
 mod entry;
 mod fork;
 mod key_set_index;
@@ -43,7 +45,7 @@ mod proof_list_index;
 mod proof_map_index;
 mod proof_map_index_next;
 mod raw_proof_map_index;
-mod readonly_fork;
+mod readonly_erased_access;
 mod temporarydb;
 mod value_set_index;