+++ exonum-java-binding/core/src/test/java/com/exonum/binding/core/blockchain/BlockchainDataIntegrationTest.java
@@ -26,6 +26,8 @@ import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.core.proxy.Cleaner;
 import com.exonum.binding.core.proxy.CloseFailuresException;
 import com.exonum.binding.core.runtime.DispatcherSchema;
+import com.exonum.binding.core.storage.database.AbstractAccess;
+import com.exonum.binding.core.storage.database.Database;
 import com.exonum.binding.core.storage.database.Fork;
 import com.exonum.binding.core.storage.database.Prefixed;
 import com.exonum.binding.core.storage.database.TemporaryDb;
@@ -34,12 +36,15 @@ import com.exonum.binding.core.storage.indices.MapIndex;
 import com.exonum.binding.core.storage.indices.ProofEntryIndexProxy;
 import com.exonum.binding.core.storage.indices.ProofMapIndexProxy;
 import com.exonum.core.messages.Runtime.InstanceState;
+import com.google.common.collect.ImmutableList;
+import java.util.Collection;
+import java.util.function.BiFunction;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
-@Disabled("ECR-4169")
 class BlockchainDataIntegrationTest {
 
   Cleaner cleaner;
@@ -75,6 +80,7 @@ class BlockchainDataIntegrationTest {
 
     Fork fork = db.createFork(cleaner);
     BlockchainData blockchainData = BlockchainData.fromRawAccess(fork, serviceName);
+
     Prefixed serviceData1 = blockchainData.getExecutingServiceData();
     // Check the service data is accessible
     IndexAddress relativeAddress = IndexAddress.valueOf(simpleIndexName);
@@ -88,9 +94,35 @@ class BlockchainDataIntegrationTest {
   }
 
   @Test
-  void getBlockchain() {
+  void createIndexProofFromForkDisallowed() throws CloseFailuresException {
+    // Setup the service data, so that the index is initialized
+    String serviceName = "test-service";
+    String simpleIndexName = "test-list";
+    try (Cleaner cl = new Cleaner()) {
+      Fork fork = db.createFork(cl);
+
+      // Use the full index name
+      String fullIndexName = serviceName + "." + simpleIndexName;
+      fork.getProofList(IndexAddress.valueOf(fullIndexName), string())
+          .add("V1");
+
+      db.merge(fork);
+    }
+
+    // Try to use a fork to create a proof for that index
     Fork fork = db.createFork(cleaner);
     BlockchainData blockchainData = BlockchainData.fromRawAccess(fork, "test-service");
+
+    Exception e = assertThrows(IllegalArgumentException.class,
+        () -> blockchainData.createIndexProof(simpleIndexName));
+    assertThat(e).hasMessageFindingMatch("(?i)(Snapshot)|(Fork)");
+  }
+
+  @ParameterizedTest
+  @MethodSource("accessConstructors")
+  void getBlockchain(BiFunction<Database, Cleaner, AbstractAccess> accessCtor) {
+    AbstractAccess access = accessCtor.apply(db, cleaner);
+    BlockchainData blockchainData = BlockchainData.fromRawAccess(access, "test-service");
 
     Blockchain blockchain = blockchainData.getBlockchain();
     // Check it works
@@ -102,10 +134,11 @@ class BlockchainDataIntegrationTest {
     assertThrows(UnsupportedOperationException.class, () -> blocks.put(blockHash, block));
   }
 
-  @Test
-  void getDispatcherSchema() {
-    Fork fork = db.createFork(cleaner);
-    BlockchainData blockchainData = BlockchainData.fromRawAccess(fork, "test-service");
+  @ParameterizedTest
+  @MethodSource("accessConstructors")
+  void getDispatcherSchema(BiFunction<Database, Cleaner, AbstractAccess> accessCtor) {
+    AbstractAccess access = accessCtor.apply(db, cleaner);
+    BlockchainData blockchainData = BlockchainData.fromRawAccess(access, "test-service");
 
     DispatcherSchema dispatcherSchema = blockchainData.getDispatcherSchema();
     // Check it works
@@ -115,5 +148,12 @@ class BlockchainDataIntegrationTest {
     // Check it is readonly
     assertThrows(UnsupportedOperationException.class,
         () -> instances.put("test-service-2", InstanceState.getDefaultInstance()));
+  }
+
+  private static Collection<BiFunction<Database, Cleaner, AbstractAccess>> accessConstructors() {
+    return ImmutableList.of(
+        Database::createFork,
+        Database::createSnapshot
+    );
   }
 }